% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{andesDAG}
\alias{andesDAG}
\title{Structure of the Andes Bayesian Network}
\format{
A 223x223 numeric matrix with row and column names given by the
network's node names. \code{andesDAG[i, j] == 1} means node \code{i} is a parent
of node \code{j}.
}
\source{
Ported from the CML package
(\url{https://github.com/stephenvsmith/CML}).
}
\usage{
data(andesDAG)
}
\description{
The 0/1 adjacency matrix of the "Andes" tutoring-system Bayesian network
(Conati et al., 1997), a 223-node, 338-edge example network used to
exercise graph-metric functions on a network much larger than \code{asiaDAG}.
}
\keyword{datasets}
