% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asiadf}
\alias{asiadf}
\title{Simulated Continuous Data from the Asia Network Structure}
\format{
A data frame with 500 rows and 8 numeric columns (\code{V1}-\code{V8},
corresponding in order to the nodes of \code{asiaDAG}).
}
\source{
Ported from the CML package
(\url{https://github.com/stephenvsmith/CML}).
}
\usage{
data(asiadf)
}
\description{
500 observations of 8 continuous (Gaussian) variables simulated from the
structure in \code{asiaDAG}, used as test data for Markov Blanket estimation.
}
\keyword{datasets}
