% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Calculate Euclidean Distance between two vectors}
\usage{
euclidean_dist(x, y)
}
\arguments{
\item{x}{NumericVector}

\item{y}{NumericVector}
}
\value{
double Euclidean distance
}
\description{
Calculate Euclidean Distance between two vectors
}
