% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslearn-package.R
\docType{package}
\name{lslearn-package}
\alias{lslearn}
\alias{lslearn-package}
\title{lslearn: Local Structure Learning for Causal DAGs}
\description{
In learning the structure of a causal DAG, users frequently want to focus on select target nodes to focus their efforts. The algorithms implemented in this package facilitate structure learning and edge orientation based on the local neighborhoods of select target nodes, as well as coordinate the information between the set of neighborhoods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenvsmith/lslearn}
  \item \url{https://stephenvsmith.github.io/lslearn/}
  \item Report bugs at \url{https://github.com/stephenvsmith/lslearn/issues}
}

}
\author{
\strong{Maintainer}: Stephen Smith \email{stephensmith13424@gmail.com}

}
\keyword{internal}
